✅ Add unit-tests for the evaluate function
