+++ tests/test_utils.py
@@ -2,7 +2,6 @@ import math
 import random
 from collections import Counter
 
-import datasets
 import pytest
 
 from kebbie.utils import (
@@ -192,42 +191,6 @@ def test_human_readable_memory(mem, s):
 )
 def test_human_readable_runtime(runtime, s):
     assert human_readable_runtime(runtime) == s
-
-
-class MockDataset:
-    def __init__(self):
-        self.n = 0
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        self.n += 1
-
-        if self.n > 20:
-            raise StopIteration
-
-        # A dummy dataset of 20 similar sentences
-        return {
-            "narrative": f"He, in fact, didn't have {self.n} wooden sticks.",
-            "dialogue": [
-                f"I have {self.n} wooden sticks.",
-                "Yes, I'm sure.",
-                "No way !",
-            ],
-        }
-
-    def shuffle(self, seed: int = 0):
-        return self
-
-
-@pytest.fixture
-def mock_load_dataset(monkeypatch):
-    # Mock dataset to avoid downloading a full-fledge dataset
-    def mock_load_dataset(*args, **kwargs):
-        return MockDataset()
-
-    monkeypatch.setattr(datasets, "load_dataset", mock_load_dataset)
 
 
 @pytest.mark.parametrize(
